refactor: move memory search to client-js
