+++ .changeset/empty-buses-attack.md
@@ -1,0 +1,5 @@
+---
+'@mastra/playground-ui': patch
+---
+
+Use client-js search memory instead of custom fetch one
