+++ .changeset/funky-lies-happen.md
@@ -1,0 +1,5 @@
+---
+'@mastra/client-js': patch
+---
+
+Add search memory in client-js
