+++ packages/playground/src/types/memory.ts
@@ -3,34 +3,3 @@ import type { MemoryConfig } from '@mastra/core/memory';
 export interface MemoryConfigResponse {
   config: MemoryConfig;
 }
-
-export interface MemorySearchResult {
-  id: string;
-  role: string;
-  content: string;
-  createdAt: string;
-  threadId?: string;
-  threadTitle?: string;
-  context?: {
-    before?: Array<{
-      id: string;
-      role: string;
-      content: string;
-      createdAt: string;
-    }>;
-    after?: Array<{
-      id: string;
-      role: string;
-      content: string;
-      createdAt: string;
-    }>;
-  };
-}
-
-export interface MemorySearchResponse {
-  results: MemorySearchResult[];
-  count: number;
-  query: string;
-  searchType?: string;
-  searchScope?: 'thread' | 'resource';
-}
